refactor: tidy schema and request
